requirements.txt
fastapi
uvicorn
mcstatus
Pillow
dnspython
cachetools
psutil
PyYAML